Qt ui code: fix qreal type problems on ARM platform
fixing problems in qt ui code comming from the fact that qreal type is float on
ARM platform and double on other platforms, which causes build errors on ARM
(not all casts are probably strictly necessary)

Bug: 12483
Reviewed-on: https://code.wireshark.org/review/15723
(cherry picked from commit b84b3f9f1d8577255a9d83c81277ebc53bcac68d)
Reviewed-on: https://code.wireshark.org/review/15724

+++ ui/qt/color_utils.cpp
@@ -99,7 +99,7 @@ const color_t ColorUtils::toColorT(const QColor color)
 
 QRgb ColorUtils::alphaBlend(const QColor &color1, const QColor &color2, qreal alpha)
 {
-    alpha = qBound(0.0, alpha, 1.0);
+    alpha = qBound(qreal(0.0), alpha, qreal(1.0));
 
     int r1 = color1.red() * alpha;
     int g1 = color1.green() * alpha;
